Use correct string conversion function for VRAM and SDMA usage
VRAM and SDMA usage can be 64-bit long numbers. Use stoull()
instead of stoi() to convert the VRAM and SDMA usage strings to
numbers.



[ROCm/rocm_smi_lib commit: fb2ed24372aa3965ef65326724acf52cf1decd35]

+++ src/rocm_smi_kfd.cc
@@ -451,7 +451,7 @@ int GetProcessInfoForPID(uint32_t pid, rsmi_process_info_t *proc,
       return EINVAL;
     }
 
-    proc->vram_usage += std::stoi(tmp);
+    proc->vram_usage += std::stoull(tmp);
 
     std::string sdma_str_path = proc_str_path;
     sdma_str_path += "/sdma_";
@@ -466,7 +466,7 @@ int GetProcessInfoForPID(uint32_t pid, rsmi_process_info_t *proc,
       return EINVAL;
     }
 
-    proc->sdma_usage += std::stoi(tmp);
+    proc->sdma_usage += std::stoull(tmp);
   }
 
   return 0;
